Rigged up ident creation.
